Added some better exception messages to the trainers.
Moved CrossProductAlphabetFactory functionality into AlphabetManager. This
had the knock-on effect of moving the two CrossProductAlphabet
implementations into the tools package.

I also replaced all the calls to CrossProductAlphabetFactory.createAlphabet
to calls to AlphabetManager.instance().createCrossProductFactory so that
everything should compile.

SimpleAlignment now has a dependancy on the tools package - mabey this is
a bad thing.


git-svn-id: file:///home/svn-repositories/biojava/biojava-live/trunk@598 7c6358e6-4a41-0410-a743-a5b2a554c398