Added DP.flatView(HMM) which returns an hmm (possibly the same object as parsed in) that it guaranteed to be flat, containing only DotState and EmissionState objects.
This allows me to make FlatModel package-private, and all the function
methods that needed flattened models now ask DP.

Also, XmlMarkovModel had been altered to use only the name parser. This
broke one of the demos. I have now attempted to fix this so that it
tries to use the most apropreate parser if available.


git-svn-id: file:///home/svn-repositories/biojava/biojava-live/trunk@614 7c6358e6-4a41-0410-a743-a5b2a554c398