Added docs to several files.
Moved the weight matrix evaluation method into DP from SingleDP so that it
is visible. There must be a better way arround this! Also, the method now
takes an offset as an additional argument so that you don't have to create
sub-lists for each place you want to try the matrix.


git-svn-id: file:///home/svn-repositories/biojava/biojava-live/trunk@674 7c6358e6-4a41-0410-a743-a5b2a554c398