Added the concept of ambiguity symbols. These simplify life greatly. Also, gap becomes the ambiguity symbol that is an empty set.
Added the concept of Distribution, that replaces the propability part of
an EmissionState. This encapsulation is more natural to work with in
practice.

The magical state stuff is simplified. Magical states have the same
alphabet as their models, and emit the gap symbol while advancing in every
direction.

all alphabet() methods in all interfaces have been migrated to
getAlphabet() to make everything more beaney.


git-svn-id: file:///home/svn-repositories/biojava/biojava-live/trunk@738 7c6358e6-4a41-0410-a743-a5b2a554c398