Fixed major bug which miscalculated the C-Terminal mass of peptides. Accidentally added two OH groups instead of two H groups if the peptide was an MH+ type.
Also changed the Constructor for MassCalc so that the user has to specify up front
whether or not it is a MH+ type of peptide. This may be a pain for some users but
 it I much safer.

added one public method

public double getTermMass()

So the user can double check that term was calculated correctly. This will be more
useful when I get C-Term and N-term mods in there.

Added on private method.

calcTermMass(String isotopicType, boolean MH_PLUS)

This just consolidates the calculation  t was performed in some other methods.
This should help me to track down errors like this in the future.


git-svn-id: file:///home/svn-repositories/biojava/biojava-live/trunk@1868 7c6358e6-4a41-0410-a743-a5b2a554c398