I have added a method: public Distribution jointDistributionOverAlignment(Alignment a,                                boolean countGaps,                                double nullWeight, int[] cols)     this just returns the joint distribution of several columns in the alignment.  It is useful for calculating mutual information for two columns in an alignment
  -  I have changed
        public Map shannonEntropy(distribution observed, double logBase)

         currently this creates a symbol->entropy map, and in the map it
puts p*log(1/p).  I think it is more natural to put log(1/p) in the map,
as this is a reflection of the uncertainty of a particular outcome (the
other is the weighted uncertainty).  I.e. if we have a weighted coin with
0.1% probability of heads, then  a head carries log(10) bits of
information.  I have also set things up so that the Map only has entries
for symbols which have non-zero probability.

   - Consequently, I have also changed

        public double bitsOfInformation(Distribution observed)

        as it reliead on shannonEntropy to calculate this.  It now
calculates the shannonEntropy map, and takes the average of the values in
this map, weighted according to the probability according to the observed
distribution.

I have also added some jUnit tests to test these methods.


git-svn-id: file:///home/svn-repositories/biojava/biojava-live/trunk@2505 7c6358e6-4a41-0410-a743-a5b2a554c398