Big hit - should have submitted this in bits.
Lots of refactoring of the vm-based ontology interpreter. Not sure if this
will survive into 1.4 any more. Added a load more ontologies - some as
tests, others with core logic that I don't want to hard code right now.

Added the .ontology.expression package that treats the knowledge as
expressions of logicals and binary predicates.


git-svn-id: file:///home/svn-repositories/biojava/biojava-live/trunk@3070 7c6358e6-4a41-0410-a743-a5b2a554c398