Added additional flexibility to fetching SCOP domains from an AtomCache.
Default behavior is unchanged. However, if the new setStrictSCOP(false) is called, the AtomCache will do its best to give a reasonable result even if the scop domain is not found. For instance, the structure of d1gbga_ would be returned for the usually invalid identifier 'd1gbg__'.
See getStructure(String) and guessScopDomain(String) for details.

git-svn-id: file:///home/svn-repositories/biojava/biojava-live/trunk@8825 7c6358e6-4a41-0410-a743-a5b2a554c398